+++ packages/react/CHANGELOG.md
@@ -9,6 +9,7 @@ description: All notable changes to this project will be documented in this file
 ### Fixed
 
 - Resolved an issue where the `SegmentGroup` component would not animate on the first click.
+- Resolved an issue whre the `Toast` component would not render custom content.
 
 ## [0.7.2] - 2023-06-30